Clean up README.md and add some performance information.
The short story is that for full builds CMake makefiles are nearly twice
as fast with 14 seconds for standard make and 8 seconds for CMake. But
for trivial builds such as tweaking main.c of an app the standard make
takes almost 7 seconds where as the CMake make take 0.3 seconds. or
about 20 times faster.

Of course your millage may very !!!!

Change-Id: I5f82a817350c7010128cd7fcc74c835fcaa14509